Reorg & comment stream
