--- defluencer/src/signature_system/mod.rs
@@ -1,4 +1,0 @@
-pub mod dag_jose;
-mod signatures;
-
-pub use signatures::{IPNSSignature, SignatureSystem};
